+++ docs/ChangeLog.md
@@ -1,6 +1,27 @@
 # Version 4
 
-## 4.14.40
+## 4.15.35
+
+### New Features
+
+`tooltips` allows custom, localized text to be included for icon tooltips
+
+### Bug Squashing
+
+fixed #1066
+
+fixed #1087 `sideBySide` properly supports `toolbarPlacement [top, bottom]`
+
+fixed #1119 
+
+fixed #1069 added input.blur()
+
+fixed #1049 fixed doc example 
+
+fixed #999 picker now looks for an element with `.input-group-addon`
+
+
+## 4.14.30
 
 ### New Features